requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
rembg>=2.0.50
google-generativeai>=0.3.0
onnxruntime>=1.16.0

